refactor(cards):  rename and reorganize queries & invalidation gestion
