+++ src/features/cards/hooks/queries/useCardsQuery.ts
@@ -1,12 +1,14 @@
-import { useInfiniteQuery } from '@tanstack/react-query';
+import { useInfiniteQuery, useQuery } from '@tanstack/react-query';
 import { httpClient } from '../../../../lib/client/http-client';
-import { CardsListResponse } from '../../types';
+import { CardBasicInfo, CardDetailedInfo, CardsListResponse } from '../../types';
 import { Filters } from '../../../filters/types';
 
 export const cardKeys = {
   all: ['card'] as const,
   list: (cardName: string, filters?: Filters) =>
     [...cardKeys.all, 'list', cardName, filters] as const,
+  detail: (id: string) => [...cardKeys.all, 'detail', id] as const,
+  detailedInfo: (id: string) => [...cardKeys.all, 'detailedInfo', id] as const,
 };
 
 export const useCards = (cardName: string, filters?: Filters) => {
@@ -43,5 +45,31 @@ export const useCards = (cardName: string, filters?: Filters) => {
     initialPageParam: 1,
     staleTime: 5 * 60 * 1000,
     retry: false,
+  });
+};
+
+export const useCardDetail = (cardId: string) => {
+  return useQuery({
+    queryKey: cardKeys.detail(cardId),
+    queryFn: async () => {
+      if (!cardId) throw new Error('Card ID is required');
+      const response = await httpClient.get<CardBasicInfo>(`/cards/${cardId}`);
+      return response;
+    },
+    enabled: !!cardId,
+    staleTime: 5 * 60 * 1000,
+  });
+};
+
+export const useCardDetailedInfo = (cardId: string) => {
+  return useQuery({
+    queryKey: cardKeys.detailedInfo(cardId),
+    queryFn: async () => {
+      if (!cardId) throw new Error('Card ID is required');
+      const response = await httpClient.get<CardDetailedInfo>(`/cards/${cardId}/details`);
+      return response;
+    },
+    enabled: !!cardId,
+    staleTime: 5 * 60 * 1000,
   });
 };
