--- src/features/sets/hooks/queries/useSetDetail.ts
@@ -1,22 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-import { httpClient } from '../../../../lib/client/http-client';
-import { SetDetailType } from '../../types';
-import { mapSetDetailData } from '../../mappers/setMapper';
-
-export const setKeys = {
-  all: ['set'] as const,
-  details: (id: string) => [...setKeys.all, 'details', id] as const,
-};
-
-export const useSetDetailedInfo = (setId: string) => {
-  return useQuery({
-    queryKey: setKeys.details(setId),
-    queryFn: async () => {
-      if (!setId) throw new Error('Set ID is required');
-      const response = await httpClient.get<SetDetailType>(`/sets/${setId}/details`);
-      return mapSetDetailData(response);
-    },
-    enabled: !!setId,
-    staleTime: 5 * 60 * 1000,
-  });
-};
